requirements.txt
arrow==1.3.0
asgiref==3.8.1
chardet==5.2.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fastkml==1.1.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pygeoif==1.5.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
reportlab==4.4.0
shapely==2.1.0
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
